requirements.txt
zipline==1.3.0
numpy==1.14
pandas==0.22.0
flake8
yahoofinancials
pyfolio
PyPortfolioOpt
scikit-learn
pytorch
pytorch-pretrained-bert==0.6.2
django==2.2.11
django-picklefield
